chore: move stand-alone methods to its file
