+++ e2e/_suites/fleet/stand-alone.go
@@ -9,12 +9,14 @@ import (
 	"strings"
 	"time"
 
+	"github.com/cenkalti/backoff/v4"
 	"github.com/elastic/e2e-testing/internal/common"
 	"github.com/elastic/e2e-testing/internal/deploy"
 	"github.com/elastic/e2e-testing/internal/installer"
 	"github.com/elastic/e2e-testing/internal/kibana"
 	"github.com/elastic/e2e-testing/internal/utils"
 	"github.com/elastic/e2e-testing/pkg/downloads"
+	"github.com/pkg/errors"
 
 	"github.com/elastic/e2e-testing/internal/elasticsearch"
 	log "github.com/sirupsen/logrus"
@@ -53,6 +55,51 @@ func (fts *FleetTestSuite) theDockerContainerIsStopped(serviceName string) error
 	}
 	fts.AgentStoppedDate = time.Now().UTC()
 
+	return nil
+}
+
+func (fts *FleetTestSuite) theStandaloneAgentIsListedInFleetWithStatus(desiredStatus string) error {
+	maxTimeout := time.Duration(utils.TimeoutFactor) * time.Minute
+	exp := utils.GetExponentialBackOff(maxTimeout)
+	retryCount := 0
+
+	agentService := deploy.NewServiceRequest(common.ElasticAgentServiceName)
+	manifest, _ := fts.getDeployer().Inspect(fts.currentContext, agentService)
+
+	waitForAgents := func() error {
+		retryCount++
+
+		agents, err := fts.kibanaClient.ListAgents(fts.currentContext)
+		if err != nil {
+			return err
+		}
+
+		if len(agents) == 0 {
+			return errors.New("No agents found")
+		}
+
+		for _, agent := range agents {
+			hostname := agent.LocalMetadata.Host.HostName
+
+			if hostname == manifest.Hostname {
+				return theAgentIsListedInFleetWithStatus(fts.currentContext, desiredStatus, hostname)
+			}
+		}
+
+		err = errors.New("Agent not found in Fleet")
+		log.WithFields(log.Fields{
+			"elapsedTime": exp.GetElapsedTime(),
+			"hostname":    manifest.Hostname,
+			"retries":     retryCount,
+		}).Warn(err)
+
+		return err
+	}
+
+	err := backoff.Retry(waitForAgents, exp)
+	if err != nil {
+		return err
+	}
 	return nil
 }
